P.I:Mongql Granular options for mutation configs

+++ server/utils/graphql/generateMutationResolvers.js
@@ -9,33 +9,35 @@ module.exports = function(resource, transformedSchema, MutationResolversTransfor
 	const capitalizedResource = resource.charAt(0).toUpperCase() + resource.substr(1);
 	const pluralizedcapitalizedResource = pluralize(capitalizedResource, 2);
 
-	const { mongql } = transformedSchema;
-	console.log(mongql.mutations);
+	const { mongql: { mutations } } = transformedSchema;
 
-	const MutationResolvers = {
-		[`create${capitalizedResource}`]: async function(parent, { data }, ctx) {
+	const MutationResolvers = {};
+	if (mutations.create[0])
+		MutationResolvers[`create${capitalizedResource}`] = async function(parent, { data }, ctx) {
 			return await createResource(ctx[capitalizedResource], ctx.user.id, data);
-		},
-		[`update${capitalizedResource}`]: async function(parent, { data, id }, ctx) {
+		};
+	if (mutations.update[0])
+		MutationResolvers[`update${capitalizedResource}`] = async function(parent, { data, id }, ctx) {
 			data.id = id;
 			return (await updateResource(ctx[capitalizedResource], [ data ], ctx.user.id, (err) => {
 				throw err;
 			}))[0];
-		},
-
-		[`update${pluralizedcapitalizedResource}`]: async function(parent, { data, ids }, ctx) {
+		};
+	if (mutations.update[1])
+		MutationResolvers[`update${pluralizedcapitalizedResource}`] = async function(parent, { data, ids }, ctx) {
 			ids.forEach((id, i) => (data[i].id = id));
 			return await updateResource(ctx[capitalizedResource], data, ctx.user.id, (err) => {
 				throw err;
 			});
-		},
-		[`delete${capitalizedResource}`]: async function(parent, { id }, ctx) {
+		};
+	if (mutations.delete[0])
+		MutationResolvers[`delete${capitalizedResource}`] = async function(parent, { id }, ctx) {
 			return (await deleteResource(ctx[capitalizedResource], [ id ], ctx.user.id))[0];
-		},
-		[`delete${pluralizedcapitalizedResource}`]: async function(parent, { ids }, ctx) {
+		};
+	if (mutations.delete[1])
+		MutationResolvers[`delete${pluralizedcapitalizedResource}`] = async function(parent, { ids }, ctx) {
 			return await deleteResource(ctx[capitalizedResource], ids, ctx.user.id);
-		}
-	};
+		};
 	let extraResolvers = {};
 	if (MutationResolversTransformers)
 		extraResolvers = MutationResolversTransformers(resource, {
